Add files via upload
A scissor can cut a paper, a rock can knock a scissor, and a paper can wrap a rock. This code randomly generates a number from 1 to 3 representing scissor, rock, and paper. Then code prompts  the user to enter a number 1, 2, or 3 and displays a message indicating
whether the user or the computer wins, loses, or draws.